initial github actions runner chart commit
Adds statefulset definition for github actions runner.

Why a statefulset?

The token provided when using "add new runner" in the UI is a temporary
registration token that can be used to register an actions runner. Once
the runner is registered (using config.sh), it stores its own
credentials locally. The registration token is short-lived (1 hour) and
cannot be renewed. But after registration is complete, the runner should
be able to continue to stay authenticated using its own credentials.

So where does the runner store credentials? `./.runner`,
`./.credentials`, etc, local to the `run.sh` and `config.sh` script.
Because they're local to the scripts, we can't mount a volume on top of
the scripts, so we copy the scripts to a volume directory and allow the
registration credentials to be written alongside the scripts inside the
volume. Thus, when the runner restarts, it doesn't need to re-register,
it simply reconnects with existing credentials.

Because of the need for persisting the registration credential and my
perceieving some importance in the hostname registered with the
credential, I've chosen a statefulset here.

Signed-off-by: Josh Bielick <jbielick@gmail.com>