Refactor ordering of StatefulSet blocks
Currently, if `persistence.enabled` is `true` and `persistence.existingClaim` is not set then Helm will throw the following error `error validating data: ValidationError(StatefulSet.spec.volumeClaimTemplates[0]): unknown field "nodeSelector" in io.k8s.api.core.v1.PersistentVolumeClaim`. This is because this combination of values causes a `volumeClaimTemplates` block to be generated before the `nodeSelector`, `affinity` and `tolerations` blocks and so these blocks are treated as attributes of the `PersistentVolumeClaim` instead of the `Pod`.